backend/requirements.txt
# Existing dependencies
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.0
google-genai>=1.0.0
pydantic>=2.0.0

# NEW: Scraper dependencies
requests>=2.31.0
Pygments>=2.17.0

# NEW: Vector store dependencies
chromadb>=0.4.22
llama-index-core>=0.10.0
llama-index-vector-stores-chroma>=0.1.0
llama-index-embeddings-gemini>=0.1.0